Create separate file for the RootSearch iterator
