+++ src/IntervalRootFinding.jl
@@ -46,6 +46,7 @@ include("krawczyk.jl")
 
 include("complex.jl")
 include("contractors.jl")
+include("rootsearch_iterator.jl")
 include("roots.jl")
 include("newton1d.jl")
 include("quadratic.jl")
